+++ Informatik/Informatik/graphics/overlay/elements/Slider.cpp
@@ -52,7 +52,7 @@ void Slider::processEvent(Menu *menu, SDL_Event e)
             bool isOver = e.button.x / SCALE_X >= x && e.button.x / SCALE_X - 10 <= x + w && e.button.y / SCALE_Y >= y && e.button.y / SCALE_Y <= y + h;
             if(isOver)
             {
-                currentValue = (float)((e.button.x / SCALE_X) - x) * (float)(maxValue - minValue) / (float)w + minValue;
+                currentValue = (int)((float)((e.button.x / SCALE_X) - x) * (float)(maxValue - minValue) / (float) w + minValue);
                 currentValue = currentValue >= maxValue ? maxValue : currentValue; // Min(current, max)
                 if(clbck != nullptr) clbck(menu, currentValue);
             }
